server/requirements.txt
flask==2.0.1
sqlalchemy==1.4.23
mysql-connector-python==8.0.26
pandas==1.3.3
scikit-learn==0.24.2
numpy==1.21.2
python-dotenv==1.0.1